Update source (beta 13)

+++ source/resolvers/hashtag.resolver.js
@@ -1,70 +1,60 @@
 import connectDB from "../database";
-import { magentaBright, redBright } from "colorette";
+import { isValidObjectId } from "mongoose";
 import Hashtag from "../models/Hashtag";
 import VTuber from "../models/VTuber";
 
 export const hashtag_resolver = {
 	Query: {
 		async hashtag() {
-			try {
-				await connectDB();
-				const data = await Hashtag.find();
-				let success = `[${magentaBright("GRAPHQL")}] Successful query!`;
-				console.log(success);
-				return data;
-			} catch (error) {
-				let data = `[${redBright("ERROR")}]\t ${error}`;
-				console.log(data);
-			}
+			await connectDB();
+			const data = await Hashtag.find();
+			return data;
 		},
 		async getHashtag(_, { ID }) {
-			try {
-				await connectDB();
-				const vtuber = await Hashtag.findById(ID);
-				if (!vtuber) throw new Error("Hashtag not found!");
-				let success = `[${magentaBright("GRAPHQL")}] Successful query!`;
-				console.log(success);
-				return vtuber;
-			} catch (error) {
-				let data = `[${redBright("ERROR")}]\t  ${error}`;
-				console.log(data);
-			}
+			await connectDB();
+			if (!isValidObjectId(ID)) throw new Error("Invalid Hashtag ID");
+			const vtuber = await Hashtag.findById(ID);
+			if (!vtuber) throw new Error("Hashtag not found!");
+			return vtuber;
 		},
 	},
 	Mutation: {
 		async addHashtag(_, { general, stream, fanart, memes }) {
 			await connectDB();
+			if (process.env.NODE_ENV === "production") {
+				throw new Error("Mutation actions are only available in development mode");
+			}
 			const new_hashtag = new Hashtag({ general, stream, fanart, memes });
 			const hashtag = await new_hashtag.save();
 			return hashtag;
 		},
 		async assignHashtagToVtuber(_, { HTID, VTID }) {
-			try {
-				await connectDB();
-				const vtuber = await VTuber.findById(VTID);
-				const hashtag = await Hashtag.findById(HTID);
-				if (!vtuber || !hashtag) {
-					throw new Error("VTuber or Hashtag not found");
-				}
-				vtuber.hashtag = hashtag._id;
-				hashtag.vtuber = vtuber._id;
-				await vtuber.save();
-				await hashtag.save();
-				return vtuber;
-			} catch (error) {
-				let data = `[${redBright("ERROR")}]\t  ${error}`;
-				console.log(data);
+			await connectDB();
+			if (process.env.NODE_ENV === "production") {
+				throw new Error("Mutation actions are only available in development mode");
 			}
+			if (!isValidObjectId(HTID) || !isValidObjectId(VTID)) {
+				throw new Error("Invalid VTuber or Hashtag ID");
+			}
+			const vtuber = await VTuber.findById(VTID);
+			const hashtag = await Hashtag.findById(HTID);
+			if (!vtuber || !hashtag) {
+				throw new Error("VTuber or Hashtag not found");
+			}
+			vtuber.hashtag = hashtag._id;
+			hashtag.vtuber = vtuber._id;
+			await vtuber.save();
+			await hashtag.save();
+			return vtuber;
 		},
 		async deleteHashtag(_, { ID }) {
-			try {
-				await connectDB();
-				const deleted = (await Hashtag.deleteOne({ _id: ID })).deletedCount;
-				return deleted;
-			} catch (error) {
-				let data = `[${redBright("ERROR")}]\t  ${error}`;
-				console.log(data);
+			await connectDB();
+			if (process.env.NODE_ENV === "production") {
+				throw new Error("Mutation actions are only available in development mode");
 			}
+			if (!isValidObjectId(ID)) throw new Error("Invalid Hashtag ID");
+			const deleted = (await Hashtag.deleteOne({ _id: ID })).deletedCount;
+			return deleted;
 		},
 	},
 };
